no poder eliminar el ciclo si tiene un aula asociada

+++ src/services/ciclosService.js
@@ -1,4 +1,5 @@
 const Ciclo = require('../models/ciclo');
+const Aula = require('../models/aula');
 
 exports.getAllCiclos = async () => {
   return await Ciclo.find();
@@ -18,6 +19,14 @@ exports.updateCiclo = async (id, data) => {
 };
 
 exports.deleteCiclo = async (id) => {
+  // Verificar si existen aulas asociadas a este ciclo
+  const aulasAsociadas = await Aula.countDocuments({ id_ciclo: id });
+  if (aulasAsociadas > 0) {
+    const err = new Error(`No se puede eliminar el ciclo porque existen ${aulasAsociadas} aula(s) asociadas.`);
+    err.statusCode = 400; // Regla de negocio
+    throw err;
+  }
+
   const result = await Ciclo.findByIdAndDelete(id);
   return !!result;
 };
